Databricks/StreamProcessing/requirements.txt
azure-eventhub